requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
msgpack==1.0.8
packaging==24.1
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.0.1
